+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/customnavview/CustomMapViewBinder.kt
@@ -11,6 +11,4 @@ class CustomMapViewBinder : MapViewBinder() {
     override fun getMapView(context: Context): MapView {
         return customMapViewFromCode(context)
     }
-
-    override val shouldLoadMapStyle: Boolean = false
 }